test(codemod): mock Clack for interactive selection; colocate tests
- interactive-selection.ts uses direct @clack/prompts imports (no injected deps).
- Add interactive-selection.test.ts beside source so vi.mock intercepts the same
  package resolution as @sku-lib/codemod; merge former tests/node coverage here.
- Keep runner delegating to interactive-selection helpers.
