+++ packages/codemod/src/transform/runner.ts
@@ -1,27 +1,20 @@
 import { glob } from 'tinyglobby';
-import {
-  intro,
-  isCancel,
-  multiselect,
-  outro,
-  path as pathPrompt,
-  select,
-  spinner,
-} from '@clack/prompts';
+import { intro, outro, spinner } from '@clack/prompts';
 import { join } from 'node:path';
-import {
-  CODEMODS,
-  JEST_TO_VITEST_STEP_SLUGS,
-  type CodemodName,
-} from '../utils/constants.js';
+import { CODEMODS } from '../utils/constants.js';
 import { Worker } from 'node:worker_threads';
 import os from 'node:os';
 import picocolors from 'picocolors';
 import debug from 'debug';
+import {
+  chooseInteractiveTransformerPaths,
+  getTargetDirectoryFromPrompt,
+  resolveCodemodModule,
+} from './interactive-selection.js';
 
 const log = debug('sku:codemod');
 
-const JEST_IMPORTS_SLUG = 'jest-to-vitest-imports' satisfies CodemodName;
+export { transformerPathsForJestSubsteps } from './interactive-selection.js';
 
 type Options = {
   dry?: boolean;
@@ -34,121 +27,6 @@ export type JobWorkerData = {
   jobs: Array<{
     filePath: string;
   }>;
-};
-
-const codemodMeta = (slug: CodemodName) =>
-  CODEMODS.find((c) => c.value === slug)?.description ?? slug;
-
-const interactiveRootOptions = () =>
-  CODEMODS.filter(
-    (c) =>
-      c.value === 'jest-to-vitest' || !c.value.startsWith('jest-to-vitest-'),
-  ).map((c) => ({
-    value: c.value,
-    label: c.value,
-    hint: c.description,
-  }));
-
-const resolveCodemodModule = (slug: CodemodName) =>
-  import.meta.resolve(`@sku-lib/codemod/codemods/${slug}`);
-
-/** Canonical-order paths for granular jest-to-vitest steps; appends imports when needed. */
-export const transformerPathsForJestSubsteps = (
-  selectedSlugs: readonly CodemodName[],
-): string[] => {
-  const selected = new Set(selectedSlugs);
-  let ordered = JEST_TO_VITEST_STEP_SLUGS.filter((s) => selected.has(s));
-
-  const needsImports =
-    [...selected].some((s) => s !== JEST_IMPORTS_SLUG) &&
-    !selected.has(JEST_IMPORTS_SLUG);
-
-  if (needsImports) {
-    ordered = [...ordered, JEST_IMPORTS_SLUG];
-  }
-
-  return ordered.map((slug) => resolveCodemodModule(slug));
-};
-
-const exitCancel = (): never => {
-  process.exit(1);
-};
-
-/** Clack prompts return `string | symbol`; cancel uses a symbol sentinel. */
-function assertClackSubmittedString(
-  value: string | symbol,
-): asserts value is string {
-  if (typeof value !== 'string') {
-    exitCancel();
-  }
-}
-
-const chooseInteractiveTransformerPaths = async (): Promise<string[]> => {
-  const rootChoice = (await select({
-    message: 'Which transform would you like to apply?',
-    options: interactiveRootOptions(),
-  })) as CodemodName;
-
-  if (isCancel(rootChoice)) {
-    exitCancel();
-  }
-
-  if (rootChoice !== 'jest-to-vitest') {
-    return [resolveCodemodModule(rootChoice)];
-  }
-
-  const pipelineMode = await select({
-    message: 'Jest → Vitest migration',
-    options: [
-      {
-        value: 'full' as const,
-        label: 'Run full pipeline',
-        hint: 'All steps in order (recommended)',
-      },
-      {
-        value: 'steps' as const,
-        label: 'Choose specific steps',
-        hint: 'Pick substeps; import synthesis runs last when needed',
-      },
-    ],
-    initialValue: 'full',
-  });
-
-  if (isCancel(pipelineMode)) {
-    exitCancel();
-  }
-
-  if (pipelineMode === 'full') {
-    return [resolveCodemodModule('jest-to-vitest')];
-  }
-
-  const picked = (await multiselect({
-    message: 'Select steps (runs in canonical order, not selection order)',
-    options: JEST_TO_VITEST_STEP_SLUGS.map((slug) => ({
-      value: slug,
-      label: slug,
-      hint: codemodMeta(slug),
-    })),
-    required: true,
-  })) as CodemodName[];
-
-  if (isCancel(picked)) {
-    exitCancel();
-  }
-
-  return transformerPathsForJestSubsteps(picked);
-};
-
-const getPathFromPrompt = async (): Promise<string> => {
-  const pathResult = await pathPrompt({
-    message: 'Which directory should the codemods run on?',
-    directory: true,
-    root: process.cwd(),
-    initialValue: '.',
-  });
-
-  assertClackSubmittedString(pathResult);
-  return pathResult;
 };
 
 export const runTransform = async (
@@ -181,7 +59,7 @@ export const runTransform = async (
     transformerPaths = await chooseInteractiveTransformerPaths();
   }
 
-  const path = _path || (await getPathFromPrompt());
+  const path = _path || (await getTargetDirectoryFromPrompt());
 
   const filesExpanded = await getAllFiles(path);
